+++ crates/remind_me_core/src/db/schema_tables.sql
@@ -148,7 +148,7 @@ CREATE TABLE IF NOT EXISTS sync_flags (
 CREATE TABLE IF NOT EXISTS sync_log (
             remote_id   TEXT NOT NULL,
             last_pull   TEXT NOT NULL DEFAULT '1970-01-01T00:00:00+00:00',
-            last_push   TEXT NOT NULL DEFAULT '1970-01-01T00:00:00+00:00', last_pull_id TEXT NOT NULL DEFAULT '', last_attempt_at TEXT NOT NULL DEFAULT '1970-01-01T00:00:00+00:00', last_push_at TEXT NOT NULL DEFAULT '1970-01-01T00:00:00+00:00', last_pull_at TEXT NOT NULL DEFAULT '1970-01-01T00:00:00+00:00',
+            last_push   TEXT NOT NULL DEFAULT '1970-01-01T00:00:00+00:00', last_pull_id TEXT NOT NULL DEFAULT '', last_attempt_at TEXT NOT NULL DEFAULT '1970-01-01T00:00:00+00:00', last_push_at TEXT NOT NULL DEFAULT '1970-01-01T00:00:00+00:00', last_pull_at TEXT NOT NULL DEFAULT '1970-01-01T00:00:00+00:00', last_pull_seq INTEGER NOT NULL DEFAULT -1,
             PRIMARY KEY (remote_id)
         );